backend/requirements.txt

djangorestframework
Django
postgres
psycopg2-pool
pytz